+++ .travis.yml
@@ -1,0 +1,24 @@
+sudo: false
+language: node_js
+node_js:
+  - 4
+  - 5
+
+# Make sure we have new NPM.
+before_install:
+  - npm install -g npm
+
+script:
+  - npm run lint
+  - npm test
+  - npm run coverage
+
+addons:
+  firefox: 'latest'
+
+before_script:
+  - export DISPLAY=:99.0
+  - sh -e /etc/init.d/xvfb start
+
+after_success:
+  - npm run coverage-publish
