+++ examples/try.js
@@ -1,2 +1,6 @@
+'use strict'
+
 var Bootstrap = require('./../src')
 var b = new Bootstrap(Bootstrap.default)
+
+console.log(b)
