+++ package.json
@@ -2,9 +2,16 @@
   "name": "libp2p-railing",
   "version": "0.2.1",
   "description": "Node.js IPFS Implementation of the railing process of a Node through a bootstrap peer list",
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "./node_modules/.bin/lab tests/*-test.js"
+    "test": "aegir-test",
+    "lint": "aegir-lint",
+    "test:node": "aegir-test node",
+    "test:browser": "aegir-test browser",
+    "build": "aegir-build",
+    "release": "aegir-release",
+    "coverage": "aegir-coverage",
+    "coverage-publish": "aegir-coverage publish"
   },
   "repository": {
     "type": "git",
@@ -23,17 +30,21 @@
   },
   "homepage": "https://github.com/diasdavid/js-ipfs-railing",
   "devDependencies": {
-    "code": "^1.4.1",
-    "lab": "^5.13.0",
+    "aegir": "^2.1.1",
+    "chai": "^3.5.0",
     "libp2p-swarm": "^0.5.0",
     "libp2p-tcp": "^0.1.1",
-    "pre-commit": "^1.1.1",
-    "standard": "^4.5.4"
+    "pre-commit": "^1.1.2"
   },
   "dependencies": {
     "ipfs-logger": "^0.1.0",
     "peer-info": "^0.3.0",
     "peer-id": "^0.3.0",
     "multiaddr": "^1.0.0"
-  }
+  },
+  "jsnext:main": "src/index.js",
+  "pre-commit": [
+    "lint",
+    "test"
+  ]
 }
